[rename🚚] 디렉터리 구조에 맞춘 경로 이동(#9)
