+++ src/main/java/com/spoteditor/backend/config/jwt/JwtConstants.java
@@ -1,0 +1,14 @@
+package com.spoteditor.backend.config.jwt;
+
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.MacAlgorithm;
+
+public class JwtConstants {
+    public static final MacAlgorithm ALGORITHM = Jwts.SIG.HS512;
+
+    public static final String TOKEN_TYPE = "JWT";
+
+    public static final String ACCESS_TOKEN = "AccessToken";
+
+    public static final String REFRESH_TOKEN = "RefreshToken";
+}
