--- src/main/java/com/spoteditor/backend/domain/user/common/dto/UserIdDto.java
@@ -1,10 +1,0 @@
-package com.spoteditor.backend.domain.user.common.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-@Data
-@AllArgsConstructor
-public class UserIdDto {
-    private Long id;
-}
