+++ src/main/java/com/spoteditor/backend/user/common/dto/UserIdDto.java
@@ -1,0 +1,10 @@
+package com.spoteditor.backend.user.common.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserIdDto {
+    private Long id;
+}
