+++ src/test/java/com/spoteditor/backend/domain/user/entity/value/OauthProviderTest.java
@@ -1,6 +1,7 @@
 package com.spoteditor.backend.domain.user.entity.value;
 
-import com.spoteditor.backend.common.exceptions.user.UserException;
+import com.spoteditor.backend.global.exception.UserException;
+import com.spoteditor.backend.user.entity.value.OauthProvider;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
